requirements.txt
absl-py==0.12.0
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
grpcio==1.38.0
idna==2.10
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.4.2
numpy==1.20.3
oauthlib==3.1.0
pathlib==1.0.1
Pillow==8.2.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
torch==1.8.1
torchvision==0.9.1
tqdm==4.61.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
